requirements.txt
requests~=2.27.1
pycryptodomex~=3.12.0
yarl~=1.7.2
tqdm~=4.62.2
